pyproject.toml
[project]
name = "steel-defect-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "torch==2.1.2+cu118",
    "torchvision==0.16.2+cu118",
    "torch==2.1.2+cu118",
    "torchvision==0.16.2+cu118",
    "ultralytics==8.0.232",
    "opencv-python==4.8.1.78",
    "numpy==1.26.4",
    "pandas==2.1.4",
    "matplotlib==3.8.2",
    "PyYAML==6.0.1",
    "tqdm==4.66.1",    
]
[tool.uv]
environments = ["sys_platform == 'win32'"]